% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-class.R
\name{rational-class}
\alias{rational-class}
\alias{rational}
\title{Rational Number Classes}
\usage{
rational(n, d, method = "R6")
}
\arguments{
\item{n}{the numerator}

\item{d}{the denominator}

\item{method}{a length = 1 character vector.  One of "R6" (default), "S3", "S4"}
}
\value{
the desired instance of the rational class
}
\description{
An S3, S4, and R6 class for a rational number
\itemize{
   \item rationalS3 - access fields with the '$' operator.
   \item rationalS4 - access fields with the '@' operator.
   \item rationalR6 - internal elements are private, so fields are accessed with accessor methods,
     \code{$getNumerator(), $getDenominator(), $getValue()}
}
The classes are designed to be used in a similar way to integers and numerics in R.

generator rational number of class \code{rationalS3}, \code{rationalS4},
and \code{rationalR6}.  Each type of class has advantages and disadvantages
in performance and flexibility
}
\section{Fields}{

\describe{
\item{\code{n,d,v}}{the numerator, denominator, and value field of the S3 class}
}}

\section{Slots}{

\describe{
\item{\code{n}}{the numerator of the S4 class}

\item{\code{d}}{the denominator of the S4 class}

\item{\code{v}}{the \code{numeric} value of the S4 class}
}}

\note{
note that \code{Inf}, \code{NA}, \code{NaN}, and \code{NULL} all fail
on is.integer() and are not permitted
}
\examples{
 a <- rational(1L, 3L, method="S3")
 stopifnot(a$n == 1L && a$d == 3L && abs(a$v - 1/3) < 1E-12)
 stopifnot(class(a) == "rationalS3")
 b <- rational(2L, 5L, method="S4")
 stopifnot(b@n == 2L && b@d == 5L && abs(b@v - 2/5) < 1E-12)
 stopifnot(class(b) == "rationalS4" && isS4(b) && is(b, "rationalS4"))
 d <- rational(3L, 7L, method="R6")
 stopifnot(d$getNumerator() == 3L && d$getDenominator() == 7L && abs(d$getValue() - 3/7) < 1E-12)
 stopifnot(class(d)[1] == "rationalR6" && is(d, "rationalR6") && is(d, "R6"))
}
