% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-compare.R
\docType{methods}
\name{rational-compare}
\alias{rational-compare}
\alias{Compare,rationalS4,numeric-method}
\alias{Compare,rationalS4,integer-method}
\alias{Compare,numeric,rationalS4-method}
\alias{Compare,integer,rationalS4-method}
\alias{Compare,rationalS4,rationalS4-method}
\alias{Ops.rationalS3}
\alias{Ops.rationalR6}
\title{Rational Number Comparisons}
\usage{
\S4method{Compare}{rationalS4,numeric}(e1, e2)

\S4method{Compare}{rationalS4,integer}(e1, e2)

\S4method{Compare}{numeric,rationalS4}(e1, e2)

\S4method{Compare}{integer,rationalS4}(e1, e2)

\S4method{Compare}{rationalS4,rationalS4}(e1, e2)

\method{Ops}{rationalS3}(e1, e2)

\method{Ops}{rationalR6}(e1, e2)
}
\arguments{
\item{e1}{rational numbers, integers, or numerics}

\item{e2}{rational numbers, integers, or numerics}
}
\description{
Rational Number Comparisons
}
\examples{
a <- rational(1L, 3L, "S3")
b <- rational(3L, 4L, "S3")
d <- 3L
e <-  20.1
stopifnot(a != b)
stopifnot(!(a == b))
stopifnot(a < b)
stopifnot(!(a > b))
stopifnot(a <= b)
stopifnot(!(a >= b))
stopifnot(a != d)
stopifnot(!(a == d))
stopifnot(a < d)
stopifnot(!(a > d))
stopifnot(a <= d)
stopifnot(!(a >= d))
stopifnot(a != e)
stopifnot(!(a == e))
stopifnot(a < e)
stopifnot(!(a > e))
stopifnot(a <= e)
stopifnot(!(a >= e))
a <- rational(1L, 3L, "R6")
b <- rational(3L, 4L, "R6")
d <- 3L
e <-  20.1
stopifnot(a != b)
stopifnot(!(a == b))
stopifnot(a < b)
stopifnot(!(a > b))
stopifnot(a <= b)
stopifnot(!(a >= b))
stopifnot(a != d)
stopifnot(!(a == d))
stopifnot(a < d)
stopifnot(!(a > d))
stopifnot(a <= d)
stopifnot(!(a >= d))
stopifnot(a != e)
stopifnot(!(a == e))
stopifnot(a < e)
stopifnot(!(a > e))
stopifnot(a <= e)
stopifnot(!(a >= e))
}
