% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-extract.R
\name{rational-extract}
\alias{rational-extract}
\alias{[,rationalS4-method}
\alias{[<-,rationalS4-method}
\alias{[[,rationalS4-method}
\alias{[[<-,rationalS4-method}
\alias{[.rationalS3}
\alias{[<-.rationalS3}
\alias{[[.rationalS3}
\alias{[[<-.rationalS3}
\alias{[.rationalR6}
\alias{[<-.rationalR6}
\alias{[[.rationalR6}
\alias{[[<-.rationalR6}
\title{Extract from a rational vector}
\usage{
\S4method{[}{rationalS4}(x, i, j, ..., drop = TRUE)

\S4method{[}{rationalS4}(x, i, j, ...) <- value

\S4method{[[}{rationalS4}(x, i, ..., drop)

\S4method{[[}{rationalS4}(x, i, j, ...) <- value

\method{[}{rationalS3}(x, i, ..., drop = TRUE)

\method{[}{rationalS3}(x, i, ...) <- value

\method{[[}{rationalS3}(x, i, ..., exact = TRUE)

\method{[[}{rationalS3}(x, i, ...) <- value

\method{[}{rationalR6}(x, i, ..., drop = TRUE)

\method{[}{rationalR6}(x, i, ...) <- value

\method{[[}{rationalR6}(x, i, ..., exact = TRUE)

\method{[[}{rationalR6}(x, i, ...) <- value
}
\arguments{
\item{x}{the rational number}

\item{i}{index specifying elements}

\item{j}{index specifying elements}

\item{...}{indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.}

\item{value}{the replacement value}

\item{exact}{controls partial matching when extracting by character}
}
\description{
Extract from a rational vector
}
\examples{
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "S4")
  stopifnot(a[2]@n == 5L)
  stopifnot(all(a[2:3]@n == c(5,6)))
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "S3")
  stopifnot(a[2]$n == 5L)
  stopifnot(all(a[2:3]$n == c(5,6)))
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "R6")
  stopifnot(a[2]$getNumerator() == 5L)
  stopifnot(all(a[2:3]$n == c(5,6)))
}
\seealso{
\code{\link{Extract}} for more full descriptions
}
