% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-functions.R
\docType{methods}
\name{rational-functions}
\alias{rational-functions}
\alias{Math,rationalS4-method}
\alias{Math2,rationalS4-method}
\alias{abs,rationalS4-method}
\alias{log,rationalS4-method}
\alias{log10,rationalS4-method}
\alias{logb,rationalS4-method}
\alias{log2,rationalS4-method}
\alias{gamma,rationalS4-method}
\alias{max,rationalS4-method}
\alias{min,rationalS4-method}
\alias{range,rationalS4-method}
\alias{prod,rationalS4-method}
\alias{sum,rationalS4-method}
\title{Functions of Rational Numbers}
\usage{
\S4method{Math}{rationalS4}(x)

\S4method{Math2}{rationalS4}(x)

\S4method{abs}{rationalS4}(x)

\S4method{log}{rationalS4}(x)

\S4method{log10}{rationalS4}(x)

\S4method{logb}{rationalS4}(x, base = exp(1))

\S4method{log2}{rationalS4}(x)

\S4method{gamma}{rationalS4}(x)

\S4method{max}{rationalS4}(x, na.rm = FALSE)

\S4method{min}{rationalS4}(x, na.rm = FALSE)

\S4method{range}{rationalS4}(x, na.rm = FALSE)

\S4method{prod}{rationalS4}(x, na.rm = FALSE)

\S4method{sum}{rationalS4}(x, na.rm = FALSE)
}
\arguments{
\item{x}{rational numbers}

\item{base}{the logarithm base in the \code{logb} function}

\item{na.rm}{will NAs be removed from a collection before the computation is performed}
}
\description{
Functions of Rational Numbers
}
\examples{
floor(rational(5L, 2L, "S4"))
sign(rational(-1L, 2L, "S4"))
round(rational(5L, 2L, "S4"), 0)
x <- abs(rational(-2L, 3L, "S4"))
stopifnot("rationalS4" \%in\% class(x))
log(rational(2L, 5L, "S4"))
log10(rational(2L, 5L, "S4"))
logb(rational(2L, 5L, "S4"), base = 5)
log2(rational(2L, 5L, "S4"))
gamma(rational(2L, 5L, "S4"))
max(rational(c(2L, 3L), c(5L, 1L), "S4"))
min(rational(c(2L, 3L), c(5L, 1L), "S4"))
range(rational(c(2L, 3L), c(5L, 1L), "S4"))
prod(rational(c(2L, 3L), c(5L, 1L), "S4"))
sum(rational(c(2L, 3L), c(5L, 1L), "S4"))
}
