% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-length.R
\name{rational-length}
\alias{rational-length}
\alias{length,rationalS4-method}
\alias{length.rationalS3}
\alias{length.rationalR6}
\alias{length.rational::rationalS7}
\title{Rational class \code{length}}
\usage{
\S4method{length}{rationalS4}(x)

\method{length}{rationalS3}(x, ...)

\method{length}{rationalR6}(x, ...)

\method{length}{`rational::rationalS7`}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rational}}

\item{...}{Other arguments passed to \code{length}}
}
\description{
Rational class \code{length}
}
\examples{
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "S4")
  stopifnot(length(a) == 3)
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "S3")
  stopifnot(length(a) == 3)
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "R6")
  stopifnot(length(a) == 3)
  a <- rational(c(3L, 5L, 6L), c(4L, 5L, 7L), "S7")
  stopifnot(length(a) == 3)
}
