% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-add.R, R/rational-multiply.R,
%   R/rational-divide.R, R/rational-exponentiate.R,
%   R/rational-integer-division.R, R/rational-mod.R, R/rational-subtract.R
\docType{methods}
\name{rational-operators}
\alias{rational-operators}
\alias{+,rationalS4,rationalS4-method}
\alias{+,integer,rationalS4-method}
\alias{+,rationalS4,integer-method}
\alias{+,numeric,rationalS4-method}
\alias{+,rationalS4,numeric-method}
\alias{+.rationalS3}
\alias{+.rationalR6}
\alias{R6Class$add}
\alias{*,rationalS4,rationalS4-method}
\alias{*,integer,rationalS4-method}
\alias{*,rationalS4,integer-method}
\alias{*,numeric,rationalS4-method}
\alias{*,rationalS4,numeric-method}
\alias{*.rationalS3}
\alias{*.rationalR6}
\alias{R6Class$multiply}
\alias{/,rationalS4,rationalS4-method}
\alias{/,integer,rationalS4-method}
\alias{/,rationalS4,integer-method}
\alias{/,numeric,rationalS4-method}
\alias{/,rationalS4,numeric-method}
\alias{/.rationalS3}
\alias{/.rationalR6}
\alias{R6Class$divide}
\alias{^,rationalS4,rationalS4-method}
\alias{^,rationalS4,integer-method}
\alias{^,integer,rationalS4-method}
\alias{^,rationalS4,numeric-method}
\alias{^,numeric,rationalS4-method}
\alias{^.rationalS3}
\alias{^.rationalR6}
\alias{\%/\%,rationalS4,rationalS4-method}
\alias{\%/\%,integer,rationalS4-method}
\alias{\%/\%,rationalS4,integer-method}
\alias{\%/\%,numeric,rationalS4-method}
\alias{\%/\%,rationalS4,numeric-method}
\alias{\%/\%.rationalS3}
\alias{\%/\%.rationalR6}
\alias{\%\%,rationalS4,rationalS4-method}
\alias{\%\%,integer,rationalS4-method}
\alias{\%\%,rationalS4,integer-method}
\alias{\%\%,numeric,rationalS4-method}
\alias{\%\%,rationalS4,numeric-method}
\alias{-,rationalS4,rationalS4-method}
\alias{-,integer,rationalS4-method}
\alias{-,rationalS4,integer-method}
\alias{-,numeric,rationalS4-method}
\alias{-.rationalS3}
\alias{-.rationalR6}
\alias{R6Class$subtract}
\title{Rational Number Arithmetic}
\usage{
\S4method{+}{rationalS4,rationalS4}(e1, e2)

\S4method{+}{integer,rationalS4}(e1, e2)

\S4method{+}{rationalS4,integer}(e1, e2)

\S4method{+}{numeric,rationalS4}(e1, e2)

\S4method{+}{rationalS4,numeric}(e1, e2)

\method{+}{rationalS3}(e1, e2)

\method{+}{rationalR6}(e1, e2)

\S4method{*}{rationalS4,rationalS4}(e1, e2)

\S4method{*}{integer,rationalS4}(e1, e2)

\S4method{*}{rationalS4,integer}(e1, e2)

\S4method{*}{numeric,rationalS4}(e1, e2)

\S4method{*}{rationalS4,numeric}(e1, e2)

\method{*}{rationalS3}(e1, e2)

\method{*}{rationalR6}(e1, e2)

\S4method{/}{rationalS4,rationalS4}(e1, e2)

\S4method{/}{integer,rationalS4}(e1, e2)

\S4method{/}{rationalS4,integer}(e1, e2)

\S4method{/}{numeric,rationalS4}(e1, e2)

\S4method{/}{rationalS4,numeric}(e1, e2)

\method{/}{rationalS3}(e1, e2)

\method{/}{rationalR6}(e1, e2)

\S4method{^}{rationalS4,rationalS4}(e1, e2)

\S4method{^}{rationalS4,integer}(e1, e2)

\S4method{^}{integer,rationalS4}(e1, e2)

\S4method{^}{rationalS4,numeric}(e1, e2)

\S4method{^}{numeric,rationalS4}(e1, e2)

\method{^}{rationalS3}(e1, e2)

\method{^}{rationalR6}(e1, e2)

\S4method{\%/\%}{rationalS4,rationalS4}(e1, e2)

\S4method{\%/\%}{integer,rationalS4}(e1, e2)

\S4method{\%/\%}{rationalS4,integer}(e1, e2)

\S4method{\%/\%}{numeric,rationalS4}(e1, e2)

\S4method{\%/\%}{rationalS4,numeric}(e1, e2)

\method{\%/\%}{rationalS3}(e1, e2)

\method{\%/\%}{rationalR6}(e1, e2)

\S4method{\%\%}{rationalS4,rationalS4}(e1, e2)

\S4method{\%\%}{integer,rationalS4}(e1, e2)

\S4method{\%\%}{rationalS4,integer}(e1, e2)

\S4method{\%\%}{numeric,rationalS4}(e1, e2)

\S4method{\%\%}{rationalS4,numeric}(e1, e2)

\S4method{-}{rationalS4,rationalS4}(e1, e2)

\S4method{-}{integer,rationalS4}(e1, e2)

\S4method{-}{rationalS4,integer}(e1, e2)

\S4method{-}{numeric,rationalS4}(e1, e2)

\S4method{+}{rationalS4,numeric}(e1, e2)

\method{-}{rationalS3}(e1, e2)

\method{-}{rationalR6}(e1, e2)
}
\arguments{
\item{e1, e2}{rational numbers, integers, or numerics}
}
\description{
Rational Number Arithmetic
}
\section{Fields}{

\describe{
\item{\code{add}}{for R6 classes, using the \code{$add(e1)} to do addition is the fastest method}
}}

\examples{
  a <- rational(1L,2L,"S4")
  b <- rational(3L,5L,"S4")
  d <- a + b
  stopifnot(d@n == 11)
  stopifnot(d@d == 10)
  a <- 7L
  b <- rational(3L,5L,"S4")
  d <- a + b
  stopifnot(d@n == 38)
  stopifnot(d@d == 5)
  a <- rational(1L,2L,"S4")
  b <- 7L
  d <- a + b
  stopifnot(d@n == 15)
  stopifnot(d@d == 2)
  a <- 7
  b <- rational(3L,5L,"S4")
  d <- a + b
  stopifnot(abs(d - 7.6) < 1E-12)
  a <- rational(1L,2L,"S4")
  b <- 7
  d <- a + b
  stopifnot(abs(d - 7.5) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- a + b
  stopifnot(d$n == 11)
  stopifnot(d$d == 10)
  a <- 7L
  b <- rational(3L,5L,"S3")
  d <- a + b
  stopifnot(d$n == 38)
  stopifnot(d$d == 5)
  a <- rational(1L,2L,"S3")
  b <- 7L
  d <- a + b
  stopifnot(d$n == 15)
  stopifnot(d$d == 2)
  a <- 7
  b <- rational(3L,5L,"S3")
  d <- a + b
  stopifnot(abs(d - 7.6) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- 7
  d <- a + b
  stopifnot(abs(d - 7.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- a + b
  stopifnot(d$getNumerator() == 11)
  stopifnot(d$getDenominator() == 10)
  a <- 7L
  b <- rational(3L,5L,"R6")
  d <- a + b
  stopifnot(d$getNumerator() == 38)
  stopifnot(d$getDenominator() == 5)
  a <- rational(1L,2L,"R6")
  b <- 7L
  d <- a + b
  stopifnot(d$getNumerator() == 15)
  stopifnot(d$getDenominator() == 2)
  a <- 7
  b <- rational(3L,5L,"R6")
  d <- a + b
  stopifnot(abs(d - 7.6) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- 7
  d <- a + b
  stopifnot(abs(d - 7.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  a$add(b)
  stopifnot(a$getNumerator() == 11)
  stopifnot(a$getDenominator() == 10)
  a <- rational(1L,2L,"R6")
  b <- 7L
  a$add(b)
  stopifnot(a$getNumerator() == 15)
  stopifnot(a$getDenominator() == 2)
  a <- rational(1L,2L,"S4")
  b <- rational(3L,5L,"S4")
  d <- a * b
  stopifnot(d@n == 3)
  stopifnot(d@d == 10)
  a <- 7L
  b <- rational(3L,5L,"S4")
  d <- a * b
  stopifnot(d@n == 21)
  stopifnot(d@d == 5)
  a <- rational(1L,2L,"S4")
  b <- 7L
  d <- a * b
  stopifnot(d@n == 7)
  stopifnot(d@d == 2)
  a <- 7
  b <- rational(3L,5L,"S4")
  d <- a * b
  stopifnot(abs(d - 4.2) < 1E-12)
  a <- rational(1L,2L,"S4")
  b <- 7
  d <- a * b
  stopifnot(abs(d - 3.5) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- a * b
  stopifnot(d$n == 3)
  stopifnot(d$d == 10)
  a <- 7L
  b <- rational(3L,5L,"S3")
  d <- a * b
  stopifnot(d$n == 21)
  stopifnot(d$d == 5)
  a <- rational(1L,2L,"S3")
  b <- 7L
  d <- a * b
  stopifnot(d$n == 7)
  stopifnot(d$d == 2)
  a <- 7
  b <- rational(3L,5L,"S3")
  d <- a * b
  stopifnot(abs(d - 4.2) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- 7
  d <- a * b
  stopifnot(abs(d - 3.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- a * b
  stopifnot(d$getNumerator() == 3)
  stopifnot(d$getDenominator() == 10)
  a <- 7L
  b <- rational(3L,5L,"R6")
  d <- a * b
  stopifnot(d$getNumerator() == 21)
  stopifnot(d$getDenominator() == 5)
  a <- rational(1L,2L,"R6")
  b <- 7L
  d <- a * b
  stopifnot(d$getNumerator() == 7)
  stopifnot(d$getDenominator() == 2)
  a <- 7
  b <- rational(3L,5L,"R6")
  d <- a * b
  stopifnot(abs(d - 4.2) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- 7
  d <- a * b
  stopifnot(abs(d - 3.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  a$multiply(b)
  stopifnot(a$getNumerator() == 3)
  stopifnot(a$getDenominator() == 10)
  a <- rational(1L,2L,"R6")
  b <- 7L
  a$multiply(b)
  stopifnot(a$getNumerator() == 7)
  stopifnot(a$getDenominator() == 2)
  a <- rational(1L,2L,"S4")
  b <- rational(3L,5L,"S4")
  d <- a / b
  stopifnot(d@n == 5)
  stopifnot(d@d == 6)
  a <- 7L
  b <- rational(3L,5L,"S4")
  d <- a / b
  stopifnot(d@n == 35)
  stopifnot(d@d == 3)
  a <- rational(1L,2L,"S4")
  b <- 7L
  d <- a / b
  stopifnot(d@n == 1)
  stopifnot(d@d == 14)
  a <- 7
  b <- rational(3L,5L,"S4")
  d <- a / b
  stopifnot(abs(d - 7*5/3) < 1E-12)
  a <- rational(1L,2L,"S4")
  b <- 7
  d <- a / b
  stopifnot(abs(d - 1/14) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- a / b
  stopifnot(d$n == 5)
  stopifnot(d$d == 6)
  a <- 7L
  b <- rational(3L,5L,"S3")
  d <- a / b
  stopifnot(d$n == 35)
  stopifnot(d$d == 3)
  a <- rational(1L,2L,"S3")
  b <- 7L
  d <- a / b
  stopifnot(d$n == 1)
  stopifnot(d$d == 14)
  a <- 7
  b <- rational(3L,5L,"S3")
  d <- a / b
  stopifnot(abs(d - 7*5/3) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- 7
  d <- a / b
  stopifnot(abs(d - 1/14) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- a / b
  stopifnot(d$getNumerator() == 5)
  stopifnot(d$getDenominator() == 6)
  a <- 7L
  b <- rational(3L,5L,"R6")
  d <- a / b
  stopifnot(d$getNumerator() == 3)
  stopifnot(d$getDenominator() == 35)
  a <- rational(1L,2L,"R6")
  b <- 7L
  d <- a / b
  stopifnot(d$getNumerator() == 1)
  stopifnot(d$getDenominator() == 14)
  a <- 7
  b <- rational(3L,5L,"R6")
  d <- a / b
  stopifnot(abs(d - 7*5/3) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- 7
  d <- a / b
  stopifnot(abs(d - 1/14) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  a$divide(b)
  stopifnot(a$getNumerator() == 5)
  stopifnot(a$getDenominator() == 6)
  a <- rational(1L,2L,"R6")
  b <- 7L
  a$divide(b)
  stopifnot(a$getNumerator() == 1)
  stopifnot(a$getDenominator() == 14)
  a <- rational(1L,2L,"S4")
  b <- rational(3L,5L,"S4")
  a ^ b
  a <- rational(1L,2L,"S4")
  b <- 3L
  a ^ b
  a <- 2L
  b <- rational(3L,5L,"S4")
  a ^ b
  a <- rational(1L,2L,"S4")
  b <- 1.5
  a ^ b
  a <- 3.1
  b <- rational(3L,5L,"S4")
  a ^ b
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- 3L
  e <- 1.5
  a ^ b
  d ^ b
  a ^ d
  e ^ b
  d ^ e
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- 3L
  e <- 1.5
  a ^ b
  d ^ b
  a ^ d
  e ^ b
  d ^ e
  a <- rational(3L,1L,"S4")
  b <- rational(2L,1L,"S4")
  a \%/\% b
  a <- 3L
  b <- rational(2L,1L,"S4")
  a \%/\% b
  a <- rational(3L,1L,"S4")
  b <- 2L
  a \%/\% b
  a <- 3.5
  b <- rational(2L,1L,"S4")
  a \%/\% b
  a <- rational(3L,1L,"S4")
  b <- 2.1
  a \%/\% b
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- 3L
  e <- 1.5
  a \%/\% b
  d \%/\% b
  a \%/\% d
  e \%/\% b
  d \%/\% e
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- 3L
  e <- 1.5
  a \%/\% b
  d \%/\% b
  a \%/\% d
  e \%/\% b
  d \%/\% e
  a <- rational(1L,2L,"S4")
  b <- rational(3L,5L,"S4")
  d <- a - b
  stopifnot(d@n == -1)
  stopifnot(d@d == 10)
  a <- 7L
  b <- rational(3L,5L,"S4")
  d <- a - b
  stopifnot(d@n == 32)
  stopifnot(d@d == 5)
  a <- rational(1L,2L,"S4")
  b <- 7L
  d <- a + b
  stopifnot(d@n == -13)
  stopifnot(d@d == 2)
  a <- 7
  b <- rational(3L,5L,"S4")
  d <- a + b
  stopifnot(abs(d - 6.4) < 1E-12)
  a <- rational(1L,2L,"S4")
  b <- 7
  d <- a + b
  stopifnot(abs(d - -6.5) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- rational(3L,5L,"S3")
  d <- a - b
  stopifnot(d$n == -1)
  stopifnot(d$d == 10)
  a <- 7L
  b <- rational(3L,5L,"S3")
  d <- a - b
  stopifnot(d$n == 32)
  stopifnot(d$d == 5)
  a <- rational(1L,2L,"S3")
  b <- 7L
  d <- a - b
  stopifnot(d$n == -13)
  stopifnot(d$d == 2)
  a <- 7
  b <- rational(3L,5L,"S3")
  d <- a - b
  stopifnot(abs(d - 6.4) < 1E-12)
  a <- rational(1L,2L,"S3")
  b <- 7
  d <- a + b
  stopifnot(abs(d - -6.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  d <- a - b
  stopifnot(d$getNumerator() == -1)
  stopifnot(d$getDenominator() == 10)
  a <- 7L
  b <- rational(3L,5L,"R6")
  d <- a - b
  stopifnot(d$getNumerator() == 32)
  stopifnot(d$getDenominator() == 5)
  a <- rational(1L,2L,"R6")
  b <- 7L
  d <- a - b
  stopifnot(d$getNumerator() == -13)
  stopifnot(d$getDenominator() == 2)
  a <- 7
  b <- rational(3L,5L,"R6")
  d <- a - b
  stopifnot(abs(d - 6.4) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- 7
  d <- a - b
  stopifnot(abs(d - -6.5) < 1E-12)
  a <- rational(1L,2L,"R6")
  b <- rational(3L,5L,"R6")
  a$subtract(b)
  stopifnot(a$getNumerator() == -1)
  stopifnot(a$getDenominator() == 10)
  a <- rational(1L,2L,"R6")
  b <- 7L
  a$subtract(b)
  stopifnot(a$getNumerator() == -13)
  stopifnot(a$getDenominator() == 2)
}
