% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-package.R
\docType{package}
\name{rational-package}
\alias{rational-package}
\title{A rational number class using multiple R class paradigms}
\description{
This package provides examples of creating rational numbers under the S3, S4, R6, and S7 object framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bertcarnell/rational}
  \item Report bugs at \url{https://github.com/bertcarnell/rational/issues}
}

}
\author{
\strong{Maintainer}: Rob Carnell \email{bertcarnell@gmail.com}

}
\keyword{internal}
