% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-print.R
\name{rational-print}
\alias{rational-print}
\alias{print,rationalS4-method}
\alias{show,rationalS4-method}
\alias{print.rationalS3}
\alias{print.rationalR6}
\title{Print an object of class Rational}
\usage{
\S4method{print}{rationalS4}(x)

\S4method{show}{rationalS4}(object)

\method{print}{rationalS3}(x, ...)

\method{print}{rationalR6}(x, ...)
}
\arguments{
\item{x}{rational object}

\item{object}{the object to show}

\item{...}{other arguments passed to print}
}
\description{
Print an object of class Rational
}
\examples{
a <- rational(1L, 3L, "S3")
print(a)
}
