% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational.R
\docType{package}
\name{rational}
\alias{rational}
\alias{rational-package}
\title{A rational number class using multiple R class paradigms}
\description{
A rational number class using multiple R class paradigms
}
