% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-group-generics.R
\name{rational_group_generics}
\alias{rational_group_generics}
\alias{S7_Math}
\alias{S7_Ops}
\alias{S7_Compare}
\alias{Math.S7_object}
\alias{find_base_generic}
\title{Ratinoal Group Generics}
\source{
\url{https://github.com/RConsortium/S7/blob/group-generics/R/method-group.R}
}
\usage{
S7_Math(x, ..., .Generic)

S7_Ops(e1, e2, ..., .Generic)

S7_Compare(e1, e2, ..., .Generic)

\method{Math}{S7_object}(x, ...)

find_base_generic(.Generic)
}
\arguments{
\item{x, z, e1, e2}{Objects used for dispatch.}

\item{..., na.rm}{Additional arguments passed to methods.}

\item{.Generic}{The name of the generic being dispatched on, i.e. if you've
  defined a method for `S7_Math` and the user calls `abs()` then `.Generic`
  will be `"abs"`.

  Use `find_base_generic()` to find the base generic that corresponds to the
  generic name.}
}
\description{
Group generics allow you to implement methods for many generics at once.
You cannot call a group generic directly; instead it is called automatically
by members of the group if a more specific method is not found. For example,
if you define a method for the `S7_Math` group generic, it will be called
when you call `abs()`, `sign()`, `sqrt()`, and many other similar generics
(see below for a complete list).
}
\details{
# Methods

The group generics contain the following methods:

* `Ops`: `r paste(group_generics_md("Compare"), group_generics_md("Arith"), group_generics_md("Logic")) `
* `Math`: `r group_generics_md("Math")`
* `Summary`: `r group_generics_md("Summary")`
* `Complex`: `r group_generics_md("Complex")`
* `matrixOps`: `r group_generics_md("matrixOps")`
}
