% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isRational.R
\name{isRational}
\alias{isRational}
\alias{is.rationalS3}
\alias{is.rationalR6}
\alias{is.rationalS4}
\alias{is.rational}
\title{Is an object from class rational

Test if an object is of class "rationalS3", "rationalS4", or "rationalR6"}
\usage{
is.rationalS3(e1, inherit = FALSE)

is.rationalR6(e1, inherit = FALSE)

is.rationalS4(e1, inherit = FALSE)

is.rational(e1, inherit = FALSE)
}
\arguments{
\item{e1}{object to be tested}

\item{inherit}{should inheritence be checked? default=\code{FALSE}}
}
\description{
Is an object from class rational

Test if an object is of class "rationalS3", "rationalS4", or "rationalR6"
}
\note{
The \code{inherit} parameter is included for speed.  There is a 
performance boost from not checking for inheritence in the class structure.
When \code{inherit==FALSE}, the the first class returned from \code{class()} is 
checked to be equal to "rationalXX".  When \code{inherit==TRUE}, then the 
\code{methods::is()} function is used.
}
\seealso{
See \code{\link{methods::is}}
}
