% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-functions.R
\name{rational-functions}
\alias{rational-functions}
\alias{Math,rationalS4-method}
\alias{Math.rationalS3}
\alias{Math.rationalR6}
\alias{Math2,rationalS4-method}
\alias{abs,rationalS4-method}
\alias{abs.rationalS3}
\alias{abs.rationalR6}
\alias{log,rationalS4-method}
\alias{log.rationalS3}
\alias{log.rationalR6}
\alias{log10,rationalS4-method}
\alias{log10.rationalS3}
\alias{log10.rationalR6}
\alias{logb,rationalS4-method}
\alias{logb.rationalS3}
\alias{logb.rationalR6}
\alias{log2,rationalS4-method}
\alias{log2.rationalS3}
\alias{log2.rationalR6}
\alias{gamma,rationalS4-method}
\alias{gamma.rationalS3}
\alias{gamma.rationalR6}
\alias{max,rationalS4-method}
\alias{max.rationalS3}
\alias{max.rationalR6}
\alias{min,rationalS4-method}
\alias{min.rationalS3}
\alias{min.rationalR6}
\alias{range,rationalS4-method}
\alias{range.rationalS3}
\alias{range.rationalR6}
\alias{prod,rationalS4-method}
\alias{prod.rationalS3}
\alias{prod.rationalR6}
\alias{sum,rationalS4-method}
\alias{sum.rationalS3}
\alias{sum.rationalR6}
\title{Functions of Rational Numbers}
\usage{
\S4method{Math}{rationalS4}(x)

\method{Math}{rationalS3}(x, ...)

\method{Math}{rationalR6}(x, ...)

\S4method{Math2}{rationalS4}(x)

\S4method{abs}{rationalS4}(x)

\method{abs}{rationalS3}(x)

\method{abs}{rationalR6}(x)

\S4method{log}{rationalS4}(x, base = exp(1))

\method{log}{rationalS3}(x, base = exp(1))

\method{log}{rationalR6}(x, base = exp(1))

\S4method{log10}{rationalS4}(x)

\method{log10}{rationalS3}(x)

\method{log10}{rationalR6}(x)

\S4method{logb}{rationalS4}(x, base = exp(1))

logb.rationalS3(x, base = exp(1))

logb.rationalR6(x, base = exp(1))

\S4method{log2}{rationalS4}(x)

\method{log2}{rationalS3}(x)

\method{log2}{rationalR6}(x)

\S4method{gamma}{rationalS4}(x)

\method{gamma}{rationalS3}(x)

\method{gamma}{rationalR6}(x)

\S4method{max}{rationalS4}(x, ..., na.rm = FALSE)

\method{max}{rationalS3}(..., na.rm = FALSE)

\method{max}{rationalR6}(..., na.rm = FALSE)

\S4method{min}{rationalS4}(x, ..., na.rm = FALSE)

\method{min}{rationalS3}(..., na.rm = FALSE)

\method{min}{rationalR6}(..., na.rm = FALSE)

\S4method{range}{rationalS4}(x, ..., na.rm = FALSE)

\method{range}{rationalS3}(..., na.rm = FALSE)

\method{range}{rationalR6}(..., na.rm = FALSE)

\S4method{prod}{rationalS4}(x, ..., na.rm = FALSE)

\method{prod}{rationalS3}(..., na.rm = FALSE)

\method{prod}{rationalR6}(..., na.rm = FALSE)

\S4method{sum}{rationalS4}(x, ..., na.rm = FALSE)

\method{sum}{rationalS3}(..., na.rm = FALSE)

\method{sum}{rationalR6}(..., na.rm = FALSE)
}
\arguments{
\item{x}{rational numbers}

\item{...}{rational numbers}

\item{base}{the logarithm base in the \code{logb} function}

\item{na.rm}{will NAs be removed from a collection before the computation is performed}
}
\description{
Functions of Rational Numbers
}
\examples{
floor(rational(5L, 2L, "S4"))
sign(rational(-1L, 2L, "S4"))
exp(rational(1L, 3L, "S3"))
exp(rational(1L, 3L, "R6"))
round(rational(5L, 2L, "S4"), 0)
x <- abs(rational(-2L, 3L, "S4"))
stopifnot("rationalS4" \%in\% class(x))
x <- abs(rational(-2L, 3L, "S3"))
stopifnot("rationalS3" \%in\% class(x))
x <- abs(rational(-2L, 3L, "R6"))
stopifnot("rationalR6" \%in\% class(x))
log(rational(2L, 5L, "S4"))
log(rational(2L, 5L, "S3"))
log(rational(2L, 5L, "R6"))
log10(rational(2L, 5L, "S4"))
log10(rational(2L, 5L, "S3"))
log10(rational(2L, 5L, "R6"))
logb(rational(2L, 5L, "S4"), base = 5)
logb(rational(2L, 5L, "S3"), base = 5)
logb(rational(2L, 5L, "R6"), base = 5)
log2(rational(2L, 5L, "S4"))
log2(rational(2L, 5L, "S3"))
log2(rational(2L, 5L, "R6"))
gamma(rational(2L, 5L, "S4"))
gamma(rational(2L, 5L, "S3"))
gamma(rational(2L, 5L, "R6"))
max(rational(c(2L, 3L), c(5L, 1L), "S4"))
max(rational(c(2L, 3L), c(5L, 1L), "S3"))
max(rational(c(2L, 3L), c(5L, 1L), "R6"))
min(rational(c(2L, 3L), c(5L, 1L), "S4"))
min(rational(c(2L, 3L), c(5L, 1L), "S3"))
min(rational(c(2L, 3L), c(5L, 1L), "R6"))
range(rational(c(2L, 3L), c(5L, 1L), "S4"))
range(rational(c(2L, 3L), c(5L, 1L), "S3"))
range(rational(c(2L, 3L), c(5L, 1L), "R6"))
prod(rational(c(2L, 3L), c(5L, 1L), "S4"))
prod(rational(c(2L, 3L), c(5L, 1L), "S3"))
prod(rational(c(2L, 3L), c(5L, 1L), "R6"))
sum(rational(c(2L, 3L), c(5L, 1L), "S4"))
sum(rational(c(2L, 3L), c(5L, 1L), "S3"))
sum(rational(c(2L, 3L), c(5L, 1L), "R6"))
}
