% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational-transformation.R
\docType{methods}
\name{rational-transformation}
\alias{rational-transformation}
\alias{as.numeric,rationalS4-method}
\alias{as.integer,rationalS4-method}
\alias{as.character,rationalS4-method}
\alias{as.rationalS4,numeric-method}
\alias{as.rationalS4,character-method}
\title{Rational Transformations}
\usage{
\S4method{as.numeric}{rationalS4}(x, ...)

\S4method{as.integer}{rationalS4}(x, ...)

\S4method{as.character}{rationalS4}(x, ...)

\S4method{as.rationalS4}{numeric}(x, cycles = 10,
  max.denominator = 2000)

\S4method{as.rationalS4}{character}(x, cycles = 10,
  max.denominator = 2000)
}
\arguments{
\item{x}{parameter to be transformed between classes}

\item{...}{additional parameters passed to underlying methods}

\item{cycles}{The maximum number of steps to be used in the continued fraction approximation process}

\item{max.denominator}{If the denominator exceeds this number, the algorithm will stop with an approximation}
}
\description{
Rational Transformations
}
\examples{
as.numeric(rational(c(2L, 3L), c(5L, 1L), "S4"))
as.integer(rational(c(2L, 3L), c(5L, 1L), "S4"))
as.integer(rational(8L, 3L, "S4"))
as.character(rational(c(2L, 3L), c(5L, 1L), "S4"))
as.rationalS4(33.3)
}
